ex4/publish/requirements.txt
py-solc-x
web3
pytest